Get rid of the `emacs-overlay`
Somehow, it prevents `nixos-rebuild`, I don't really know why, maybe an
incompatibility between `nix-doom-emacs`'s expectations of its
`emacs-overlay` input and what we actually give it. For now, we'll go
back to a regular Emacs 28.2. That's a shame, but at least it will work
better and `nixos-rebuild` should be faster.