Use gcr to provide the SSH agent
This partially fixes the SSH agent mess, but it isn't 100% great just
yet. We hit yet again the same problem, which is that we cannot use the
value of `$XDG_RUNTIME_DIR` in the configuration. By putting things in
initExtra, everything will work well in a Shell. However, a standalone
graphical Emacs will currently not pick up on the agent.